apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.18
# Auth
pyjwt>=2.8.0
bcrypt>=4.0.0
email-validator>=2.1.0
# AI Video Analysis
librosa>=0.10.2
soundfile>=0.12.1
anthropic>=0.45.0
numpy>=1.24.0
